+++ data-pipeline/tests/unit/aisr/test_query.py
@@ -4,38 +4,10 @@ Tests for interacting with AISR
 
 # pylint: disable=missing-function-docstring
 
-from unittest.mock import Mock
-
 import requests
-from data_pipeline.aisr.authenticate import (
-    _get_access_token_using_response_code,
-    _get_code_from_response,
-)
 from data_pipeline.aisr.query import S3UploadHeaders, _get_put_url, put_file_to_s3
 
 UPLOAD_FILE_NAME = "test_file.csv"
-
-
-def test_request_access_token_with_code(fastapi_server):
-    test_realm_url = f"{fastapi_server}/auth/realms/idepc-aisr-realm"
-
-    with requests.Session() as local_session:
-        token = _get_access_token_using_response_code(
-            local_session, test_realm_url, "test_code"
-        )
-
-    assert token == "mocked-access-token", "Access token should be returned"
-
-
-def test_extract_code_from_auth_response_headers(fastapi_server):
-    test_realm_url = f"{fastapi_server}/auth/realms/idepc-aisr-realm"
-    mock_response = Mock()
-    mock_response.status_code = 302
-    mock_response.headers = {"Location": f"{test_realm_url}#code=test_code"}
-
-    code = _get_code_from_response(mock_response)
-
-    assert code == "test_code", "Code should be extracted from the Location header"
 
 
 def test_can_get_put_url(fastapi_server):
